requirements.txt
Django==4.2.5
matplotlib==3.7.2
python_dateutil==2.8.2
Requests==2.31.0
